Refactor incremental sql conversion logic and cleanup
