+++ coral-incremental/build.gradle
@@ -2,11 +2,9 @@ dependencies {
   compile project(':coral-hive')
 
   testCompile(deps.'hive'.'hive-exec-core') {
-    exclude group: 'org.apache.avro', module: 'avro-tools'
-    // These exclusions are required to prevent duplicate classes since we include
-    // shaded jar above
+    // This exclusion is required to prevent duplicate classes since we include
+    // shaded 'calcite-core' transitively
     exclude group: 'org.apache.calcite', module: 'calcite-core'
-    exclude group: 'org.apache.calcite', module: 'calcite-avatica'
   }
 
   testCompile deps.'hadoop'.'hadoop-mapreduce-client-core'
