+++ coral-incremental/src/test/java/com/linkedin/coral/incremental/RelToIncrementalSqlConverterTest.java
@@ -9,6 +9,7 @@ import java.io.File;
 import java.io.IOException;
 
 import org.apache.calcite.rel.RelNode;
+import org.apache.calcite.sql.SqlNode;
 import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.api.MetaException;
@@ -16,6 +17,8 @@ import org.apache.hadoop.hive.ql.metadata.HiveException;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+
+import com.linkedin.coral.transformers.CoralRelToSqlNodeConverter;
 
 import static com.linkedin.coral.incremental.TestUtils.*;
 import static org.apache.calcite.sql.type.OperandTypes.*;
@@ -37,10 +40,16 @@ public class RelToIncrementalSqlConverterTest {
     FileUtils.deleteDirectory(new File(conf.get(CORAL_INCREMENTAL_TEST_DIR)));
   }
 
+  public String convert(RelNode relNode) {
+    RelNode incrementalRelNode = RelNodeIncrementalTransformer.convertRelIncremental(relNode);
+    CoralRelToSqlNodeConverter converter = new CoralRelToSqlNodeConverter();
+    SqlNode sqlNode = converter.convert(incrementalRelNode);
+    return sqlNode.toSqlString(converter.INSTANCE).getSql();
+  }
+
   public String getIncrementalModification(String sql) {
-    RelToIncrementalSqlConverter converter = new RelToIncrementalSqlConverter();
     RelNode originalRelNode = hiveToRelConverter.convertSql(sql);
-    return converter.convert(originalRelNode);
+    return convert(originalRelNode);
   }
 
   @Test
@@ -114,6 +123,24 @@ public class RelToIncrementalSqlConverterTest {
         "SELECT t1.x, t1.y\n" + "FROM (SELECT t.x, t.y\n" + "FROM (SELECT *\n" + "FROM test.bar1_delta AS bar1_delta\n"
             + "UNION ALL\n" + "SELECT *\n" + "FROM test.bar2_delta AS bar2_delta) AS t\n" + "GROUP BY t.x, t.y\n"
             + "UNION ALL\n" + "SELECT *\n" + "FROM test.bar3_delta AS bar3_delta) AS t1\n" + "GROUP BY t1.x, t1.y";
+    assertEquals(getIncrementalModification(sql), expected);
+  }
+
+  @Test
+  public void testSelectSpecific() {
+    String sql = "SELECT a FROM test.foo";
+    String expected = "SELECT foo_delta.a\n" + "FROM test.foo_delta AS foo_delta";
+    assertEquals(getIncrementalModification(sql), expected);
+  }
+
+  @Test
+  public void testSelectSpecificJoin() {
+    String sql = "SELECT test.bar2.y FROM test.bar1 JOIN test.bar2 ON test.bar1.x = test.bar2.x";
+    String expected = "SELECT t0.y0 AS y\n" + "FROM (SELECT *\n" + "FROM (SELECT *\n" + "FROM test.bar1 AS bar1\n"
+        + "INNER JOIN test.bar2_delta AS bar2_delta ON bar1.x = bar2_delta.x\n" + "UNION ALL\n" + "SELECT *\n"
+        + "FROM test.bar1_delta AS bar1_delta\n" + "INNER JOIN test.bar2 AS bar2 ON bar1_delta.x = bar2.x) AS t\n"
+        + "UNION ALL\n" + "SELECT *\n" + "FROM test.bar1_delta AS bar1_delta0\n"
+        + "INNER JOIN test.bar2_delta AS bar2_delta0 ON bar1_delta0.x = bar2_delta0.x) AS t0";
     assertEquals(getIncrementalModification(sql), expected);
   }
 }
